Organize ingress into v1 and v2 packages
The network reader package is now ingress/v1. What was once just
ingress is now ingress/v2.
