+++ src/metron/ingress/v1/mock_metric_batcher_test.go
@@ -1,4 +1,4 @@
-package networkreader_test
+package ingress_test
 
 type mockMetricBatcher struct {
 	BatchIncrementCounterCalled chan bool
