--- src/metron/writers/byte_array_writer.go
@@ -1,5 +1,0 @@
-package writers
-
-type ByteArrayWriter interface {
-	Write(message []byte)
-}
